refactor : Results 클래스의 getSummary 함수 가독성 있게 수정

+++ src/main/java/lotto/view/OutputView.java
@@ -22,7 +22,7 @@ public class OutputView {
     }
 
     public static void printResults(final Results results) {
-        TreeMap<WinningInfo, Long> summary = new TreeMap<>(results.getSummary());
+        TreeMap<WinningInfo, Integer> summary = new TreeMap<>(results.getSummary());
         Iterator<WinningInfo> iteratorKey = summary.keySet()
                 .iterator();
